Refactor: move online inspector to its own module
